Use getPixel:atX:y: directly

+++ src/color.rs
@@ -4,9 +4,9 @@
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Rgb {
     /// The amount of red, ranging from 0.0 to 1.0.
-    pub red: f64,
+    pub red: u8,
     /// The amount of green, ranging from 0.0 to 1.0.
-    pub green: f64,
+    pub green: u8,
     /// The amount of green, ranging from 0.0 to 1.0.
-    pub blue: f64,
+    pub blue: u8,
 }
